[WIP] : 캐릭터 선택 및 닉네임 설정 구현 중
