+++ Project/Assets/03.Code/Scripts/LogInSystem/CharaterSelect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a71fb8de08dd7f1478e0105ad64b7dbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
