+++ Project/Assets/03.Code/Scripts/UI/Page.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Page : MonoBehaviour
+{
+
+}
