+++ Project/Assets/03.Code/Scripts/UI/Page.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cf925d004dcde554c8fdc15fd02cd85a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
