+++ Project/Assets/03.Code/Scripts/UI/PageManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7ba47b844dd8f841a4854dacc928b4f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
